Use size_t for gl_pathc as asked in the libc's Makefile: # If you bump SHLIB_MAJOR, see standards/55112.
PR:             55112
Reviewed by:    ume

+++ include/glob.h
@@ -44,7 +44,7 @@
 
 struct stat;
 typedef struct {
-	int gl_pathc;		/* Count of total paths so far. */
+	size_t gl_pathc;	/* Count of total paths so far. */
 	int gl_matchc;		/* Count of paths matching pattern. */
 	int gl_offs;		/* Reserved at beginning of gl_pathv. */
 	int gl_flags;		/* Copy of flags parameter to glob. */
